Day 2

+++ demo-helloworld/src/Exercise20241210.java
@@ -1,3 +1,4 @@
+
 public class Exercise20241210 {
   public static void main(String[] args) {
     double priceForApple = 7.3;
@@ -45,7 +46,7 @@ public class Exercise20241210 {
 
     byte maxByte = 120;
     byte minByte = -128;
-    byte miniByte = maxByte + minByte;
+    byte miniByte = (byte) (maxByte + minByte);
     // Step1: byte value + int value => int value
     // Step2: cannot assign int value to a byte variable
     System.out.println(miniByte);
